% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(dat,n.rep=10,ind=1,type="ordinary")
  
  plotBootstrap(d.boot,alpha=1/2)
}
\arguments{
\item{n.rep}{number of replicates}

\item{ind}{the index of the column to be analyzed}

\item{dat}{data to be passed into bootstrap function, it needs to be in a data.frame format}

\item{type}{the type of resampling, can be "ordinary" (the default), "parametric", "balanced", "permutation", or "antithetic". See boot::boot parameter sim for details.}

\item{d.boot}{bootstrapped data}

\item{alpha}{transparency adjustment, between 0 to 1.}
}
\value{
\itemize{ 

\item{Inidvidualbootstrap} returns a matrix of bootstrapped samples, 
each row is a bootstrapped replicate of original dataset, and the length of columns are matching the size of original dataset. for example, original data has 40 data points and you want to bootstrap 10 such datasets, you will have a 10 x 40 (row x col) datasets. 
}
}
\description{
Bootstrap confidience intervals with standard errors. bootstrap resamples dataset (e.g. diffusion coefficients) to calculate confidience intervals for a statistic measure of dataset.
}
\details{
A wrapper of boot::boot function adapted for data format in sojourner package.
}
\examples{
# read in using readDiatrack
folder=system.file("extdata","SWR1",package="sojourner")
trackll=readDiatrack(folder)

# filter track based length
trackll.flt=filterTrack(trackll,filter=c(min=5,max=Inf))
MSD=msd(trackll.flt,dt=6,summarize=TRUE,plot=TRUE)
dcoef=Dcoef(MSD=MSD,method="static",plot=TRUE)

# bootstrap new datasets
d.boot=bootstrap(dcoef)

# plot bootstrapped data
plotBootstrap(d.boot)
}
