% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTrackll.R
\docType{methods}
\name{combineTrackll}
\alias{combineTrackll}
\title{combineTrackll}
\usage{
combineTrackll(trackll=c(trackll1,trackll2),merged=T)
}
\arguments{
\item{merged}{An Logical indicate if the tracklls to combine are merged or not.}

\item{total}{The number of tracklls to combine.}
}
\value{
\itemize{
\item{trackll:} combined trackll.
}
}
\description{
Combine multiple tracklls into one trackll.
}
\details{
Combine multiple track listw (tracklls) from multiple folders into one trackll, i.e. combining track infomation
         from files in multiple folders (replicates) together as if they are in one folder. The tracklls can be
         either merged or un-merged.
         
         If the combined tracklls are merged, users will be prompted to input a combined attribute name for the combined trackll.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, merge tracks from multiple files.
folder1=system.file("extdata","HSF",package="smt")
trackll1=createTrackll(interact=F,folder1,input=2, cores = 2)
trackll1=maskTracks(folder1,trackll1)
trackll1=mergeTracks(folder1,trackll1)

folder2=system.file("extdata","HSF_2",package="smt")
trackll2=createTrackll(interact=F,folder2,input=2, cores = 2)
trackll2=maskTracks(folder2,trackll2)
trackll2=mergeTracks(folder2,trackll2)

# Combine the tracklls together, input trackll names when prompted,
trackll=combineTrackll(trackll=c(trackll1,trackll2),merged=T)
}
