% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFolder.R
\docType{methods}
\name{compareFolder}
\alias{compareFolder}
\title{compareFolder}
\usage{
compareFolder(folders=c(folder1,folder2,...),input=1,ab.track=F,cores=1)
}
\arguments{
\item{folders}{a vector storing paths to the folders location."..."
indicates multiple (unlimited) folders can be added into the function.}

\item{input}{Input file type (Diatrack .txt file = 1; Diatrack .mat session
file = 2; ImageJ .csv file = 3; SlimFast .txt file = 4).}

\item{ab.track}{a Logical indicating if absolute coordinates should be used.}

\item{cores}{Number of cores used for parallel computation. This can be the
cores on a workstation, or on a cluster.}
}
\value{
\itemize{
\item{trackll} A list of tracks, each item of a track list correspond to a
folder. This list can then be fed into other functions for comparison. }
}
\description{
compare folders with Diatrack output files. merge track files
  in each folder into one item of a track list. This list can then be fed
  into other functions for comparison. It keeps folder information as names
  of the resulting list.
}
\examples{
folder1=system.file("extdata","SWR1",package="sojourner")
folder2=system.file("extdata","HTZ1",package="sojourner")
trackll=compareFolder(c(folder1,folder2))
str(trackll,max.level=1)
}
