% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dcoef.R
\docType{methods}
\name{Dcoef}
\alias{Dcoef}
\title{Dcoef}
\usage{
Dcoef( MSD=NULL,trackll=NULL,dt=6,filter=c(min=7,max=Inf),rsquare=0.8,
       resolution=0.107,binwidth=NULL,method=c("static","percentage",
       "rolling.window"),plot=FALSE,output=FALSE,t.interval=0.01,
       profile=NULL)
}
\arguments{
\item{MSD}{Mean Square Displacement calculated using msd() function. 
Either MSD or trackll can be passed into Dcoef for calculation of diffusion
coefficient.}

\item{trackll}{Track list output from readDiatrack().}

\item{dt}{Time intervals. Default 6.}

\item{filter}{a vector specifies the minimum and max length of trajecotries
to be analyzed. Take only trajectories that has number of frames greater
than (>=) min and less than (<) max.}

\item{rsquare}{rsquare filter on Dcoef results. Default to be 0.8. Set 
value to 0 if rsquare filter is not desired.}

\item{resolution}{ratio of pixel to uM.}

\item{binwidth}{binwidth used for histogram. Default NULL, automatically
assign binwidth.}

\item{method}{"static", uses time lags 2~5 to calculate diffusion
coefficient; "percentage", uses (tierd) percentage (default 0.25) of time
lags (see Details). "rolling.window", time lags uses for Dcoef follows a
rolling window with specified window size (default 4).}

\item{plot}{A parameter for plotting. Default FALSE, no plot; If TRUE,
automatically plots 'histogram' with count information, binwidth can be
set through parameter binwidth; as well as 'density' with
density/frequency.}

\item{output}{An Logical indicate if output should be generated. See Values
for detail.}

\item{t.interval}{time interval between frames, default 0.010 s (10ms).}

\item{profile}{Location of preference file. By default (NULL), it is stored
at : system.file("extdata","PREF","profile.csv",package="sojourner").
User can provide preference file by specifying the location of the file, 
e.g. profile="/Users/shengliu/Desktop/profile.csv".}
}
\value{
\itemize{
\item \emph{Dcoef} A list of Dcoef for each file in trackll.
\item \emph{PDF} Log.Dcoef histogram fitted with density curve,
 when plot = TRUE.
\item \emph{csv} Dcoef output in csv format, when output = TRUE.
}
}
\description{
Caclulate diffusion coefficient (Dcoef) for trajecotries.
}
\details{
Generic parameters (parameter applied to all methods, such as
  resolution etc) are set in the function. Method dependent parameters (such
  as lag.start, lag.end for method = "static"), are stored in profile.csv in
  PREF folder under extdata. To change preference parameter, can either
  programably or manually go to folder
  system.file("extdata","PREF","profile.csv",package="sojourner"), and c
  hange the profile.csv.

lag.start: time lag used as start of dt for compute Dcoef. Default 2.

lag.end: Time lag used as end of dt for compute Dcoef. Default 2.

method for calculating Dcoef:
\itemize{
    \item \bold{static} stabilize the number of time lags used for fitting
    using time lag 2~ 5 despite the total time lags measured.
    \item \bold{percentage} "percentage", uses (tierd) percentage (default
    0.25) of time lags.
\tabular{rlll}{
    [,1] \tab TrackLength \tab Percentage \tab TimeLagsForFitting\cr
    [,2] \tab 31~ \tab 0.25 \tab 2~5-2~7\cr
    [,3] \tab 22~30 \tab 0.25 \tab 2~5-2~7\cr
    [,4] \tab 15~21 \tab 0.4 \tab 2~5-2~7\cr
    [,5] \tab 10~15 \tab 0.6 \tab 2~5-2~7\cr
    [,6] \tab 7~9 \tab 1 \tab 2~5-2~7
}

    \item \bold{rolling.window}  time lags uses for Dcoef follows a rolling
    window with specified window size (default 4).
}
}
\examples{
# compare files
folder=system.file("extdata",'SWR1',package="sojourner")
trackll = createTrackll(folder=folder, input=3)
MSD=msd(trackll=trackll)
Dcoef(MSD=MSD,method="static",plot=TRUE)

# compare folders
folder1=system.file("extdata",'SWR1',package="sojourner")
folder2=system.file("extdata",'HTZ1',package="sojourner")
trackll2=compareFolder(folders=c(folder1,folder2), input=3)
Dcoef(trackll=trackll2,method="percentage",plot=TRUE)
Dcoef(trackll=trackll2,method="rolling.window",plot=TRUE)
}
