% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionMap.R
\docType{methods}
\name{diffusionMap}
\alias{diffusionMap}
\title{diffusionMap}
\usage{
diffusionMap(filename)
}
\arguments{
\item{filename}{trackll file path in .csv}
}
\description{
Take in a .csv trackll file and perform diffusion map analysis.
}
\details{

}
\examples{

}
