% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTrack.R
\docType{methods}
\name{filterTrack}
\alias{filterTrack}
\title{filterTrack}
\usage{
filterTrack(trackll,filter=c(min=7,max=Inf)))
trimTrack(trackll,trimmer=c(min=1,max=32)))
}
\arguments{
\item{trackll}{Track list output from readDiatrack().}

\item{filter}{length of track used for filtration. Only tracks pass through filter will be selected.}

\item{trimmer}{length of track used for trimming. All tracks will be trimmed by the trimmer.}
}
\value{
\itemize{
\item{trackll} filtered or trimmed tracks.
}
}
\description{
methods for filter and trim tracks based on track length.
}
\details{
filter is used to filter out tracks that has length within a
  specified range (default 5~Inf); On the other hand, despite the lengths of
  tracks, trimmer is used to trim /cutoff all tracks to a specified range
  (default 1~32).
}
\examples{
folder=system.file("extdata","SWR1",package="sojourner")
trackll=readDiatrack(folder)

trackll.filter=filterTrack(trackll,filter=c(7,Inf))
trackll.trim=trimTrack(trackll,trimmer=c(1,20))

# see the min and max length of the trackll
# trackLength() is a helper function output track length of trackll
lapply(trackLength(trackll),min)
lapply(trackLength(trackll.filter),min)

lapply(trackLength(trackll),max)
lapply(trackLength(trackll.trim),max)

}
