% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNormDistr.manual.R
\docType{methods}
\name{fitNormDistr.manual}
\alias{fitNormDistr.manual}
\title{fitNormDistr.manual}
\usage{
fitNormDistr.manual(data, meanRange=NULL, proportion=NULL, means=NULL, 
sd=NULL, log.transform=F, output=F, components=NULL, binwidth=NULL, seed=NULL)
}
\arguments{
\item{data}{single output from diffusion coefficient calculated from Dcoef(), like dcoef[[1]].}

\item{meanRange}{list of vectors, each of which have the min&max that estimates a range where a peak of the distribution should lie.}

\item{proportion}{numeric vector with estimates of each component's proportion of the whole data.}

\item{means}{numeric vector with estimates of mean(mu) values for each component.}

\item{sd}{numeric vector with estimates of standard deviation(sigma) values for each component.}

\item{log.transform}{logical indicate if log10 transformation is needed, default F.}

\item{output}{Logical indicaring if output file should be generated.}

\item{components}{parameter specifying the number of components to fit. If NULL (default), a components analysis is done to determine the most likely components and this number is then used for subsequent analysis.}

\item{binwidth}{binwidth for the combined plot. If NULL (default), will automatic assign binwidth.}

\item{seed}{Seed for random number generator. This makes each run easily repeatable. Seed will be automatically assigned if no seed is specified (default). The seed information is stored as an attribute of the returned object. The seed can also be output to a txt file when output=T.}
}
\value{
a single element from the list result that would result from FitNormDistr().
\describe{
\item{proportions}{The proportions of mixing components.}
\item{mean}{The Means of the components.}
\item{sd}{The Standard Deviations (SD) of components if not log transformed; if log transformed, it is then interpreted as Coefficient of Variation (CV).}
\item{loglik}{The log likelihood, useful for compare different fitting result, the bigger the better fit.}
}
}
\description{
fit a normal distribution to diffusion coefficient caclulated with specified parameters
}
\details{
components analysis uses the likelihood ratio test (LRT) to assess the number of mixture components. This will be used if there are no estimates/information
regarding the data are provided. One should provide just the meanRange or the means, sd, and proportion. If both are provided, the function may raise an
error or run using the meanRange and ignoring other parameters.
}
\examples{

# compare folders
folder1=system.file("extdata","SWR1",package="sojourner")
folder2=system.file("extdata","HTZ1",package="sojourner")
trackll=compareFolder(c(folder1,folder2))
MSD=msd(trackll=trackll)
dcoef=Dcoef(MSD,dt=6,plot=T,output=F)
# fit first element of dcoef (SWR1)
a=fitNormDistr.manual(dcoef[[1]],components=NULL,log.transform=F,output=F)
# to repeat a
b=fitNormDistr.manual(dcoef[[2]],components=NULL,log.transform=F,output=F,seed=attr(a,"seed"))
# provide possible mean ranges assuming 2 component fit
c=fitNormDistr.manual(dcoef[[2]],meanRange=list(c(0.2,0.4),c(0.6,0.8)),log.transform=F,output=F)

}
