% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitResidenceTime.R
\docType{methods}
\name{fitRT}
\alias{.one.comp.fit.rt}
\alias{fitRT}
\title{Fitting Residence time (1-CDF)}
\usage{
fitRT(trackll=NULL,x.max=30,N.min=1.5,t.interval=0.5)
}
\arguments{
\item{t.interval}{time interval for image aquisition. Default 0.5 sec.}

\item{trackll}{trajectory list generated by createTrackll() and processing. if NULL, user will be prompted to enter the trackll name.}

\item{x.max}{The maximum range of X axis, i.e. time, for the output plot. Default 30 sec.}

\item{N.min}{The minimal duration of trajectory length in the unit of second, trajectories shorter than N.min will be filtered out. Default 1.5 sec.}
}
\value{
\itemize{
\item{On the Console output:} Result and parameters of goodness of the fit.
\item{Plot:} fitting curves will be plotted over the raw data, with the result of two-component fitting shown in the plot area.
}
}
\description{
Caclulate average residence time for trajecotries by
             fitting 1-CDF (survival distribution).
}
\details{
Calculating average residence time of particles by fitting 1-CDF (survival distribution) of its trajectorys.

Upon running of the function, users will be prompted to input the name of the track list (trackll).
Input processed, merged trackll and the fitting will start right away.
The maximum time range to be plotted can be set using x.max, this will not change the fitting result.
Fitting result is determined by the input trackll, and N.min (filtering of the trackll).
}
\examples{
# Generate trackll, and process,
# e.g. mask region of interest, merge tracks from multiple files.
folder=system.file("extdata","SWR1",package="smt")
trackll=createTrackll(interact=F,folder,input=1)
trackll=maskTracks(folder,trackll)
trackll=mergeTracks(folder,trackll)

# Fit the residence time of trackll
fitRT(trackll=NULL,x.max=30,N.min=1.5,t.interval=0.5)
}

