% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\docType{methods}
\name{getCI}
\alias{getCI}
\title{getCI}
\usage{
getCI(samples,confidence=0.95,output=F)
}
\arguments{
\item{samples}{diffusion coefficient calculated from Dcoef().}

\item{confidence}{the level of confidence that is used to calculate the confidence interval.}

\item{output}{Logical indicating if output file should be generated.}
}
\value{
list of items, each of which will contain:
\describe{
\item{Estimate}{Mean estimated from sample}
\item{CI lower}{Lower bound of the confidence interval}
\item{CI upper}{Upper bound of the confidence interval}
\item{Std. Error}{Std. Error for given data}
}
}
\description{
given number of samples, return Confidence Interval information(mean, bounds, std err) for diffusion
coefficient and the proportions of each components.
}
\details{
Supplied with number of samples as dcoef, uses that to calculate the confidence range.
Each entry in dcoeflist will consist of number of samples regarding that data.
May hit an error for bootstrapping due to random number generation, but re-running the function will do the job.
}
\examples{

# compare folders
folder1=system.file("extdata","SWR1",package="sojourner")
folder2=system.file("extdata","HTZ1",package="sojourner")
trackll=compareFolder(c(folder1,folder2))
#get msd
MSD=msd(trackll=trackll)
#run Dcoef()
dcoef=Dcoef(MSD,dt=6,plot=T,output=F)
#resample data with bootstrapping(get 10 resampled data)
bootstraps=bootstrap(dcoef, attribute="slope")
# get confidence interval for this bootstrapped samples
a=getCI(bootstraps)
# to manually set confidence to 80\% and generate output csv file
b=getCI(bootstraps, confidence=0.8, output=T)

}
