% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmTrackll.R
\docType{methods}
\name{hmmTrackll}
\alias{hmmTrackll}
\title{hmmTrackll}
\usage{
hmmTrackll(trackll)
}
\arguments{
\item{trackll}{list of trajectorise created by crreateTrackll().}
}
\description{
Convert trackll to a (bivarte) displacement format for Hidden Markove Model fitting.
}
\details{
dx, displacement on x.

dy, displacement on y. 

time, lenght of time a particle has being traveling. 

trackStepIndex, steps index within a track.
}
\examples{
folder=system.file("extdata","SWR1",package="sojourner"
trackll=createTrackll(folder=folder)
hTrackll=hmmTrackll(trackll)

}
