% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTracks.R
\docType{methods}
\name{mergeTracks}
\alias{mergeTracks}
\title{mergeTracks}
\usage{
mergeTracks(folder, trackll)
}
\arguments{
\item{folder}{Full path to the output files.}

\item{trackll}{A list of track lists.}
}
\value{
modified trackll
}
\description{
merge track lists in a list into one
}
\details{
IMPORTANT: Once a trackll has been merged, it cannot be masked using maskTracks().

Merging creates a list list of data.frames. 
First level is the folder name, second level is a list of data.frames/tracks from all output files merged into one.

If not merged, track lists takes the name of individual files in the folder.
If merged, the single merged track list takes the folder name.
}
\examples{
#Basic masking with folder path with image masks
folder = system.file("extdata", "SWR1", package = "sojourner")
trackll = readDiatrack(folder)
trackll.merged <- mergeTracks(folder = folder, trackll = trackll)
}
