% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividual_Tracks.R
\docType{methods}
\name{plotIndividual_Tracks}
\alias{plotIndividual_Tracks}
\title{plotIndividual_Tracks}
\usage{
<<<<<<< HEAD:man/plotIndividual_Tracks-methods.Rd
plotIndividual_Tracks(trackll=trackll,grid.size=c(1000,1000),
resolution=0.107,t.interval=0.5)
=======
plotIndividualTracks(trackll=trackll,grid.size=c(1000,1000),resolution=0.107,t.interval=0.5,sort="RT",legend="Both",lag=c(2:5),dt=6)
>>>>>>> develop:man/plotIndividualTracks-methods.Rd
}
\arguments{
\item{trackll}{trajectory list generated by createTrackll() and 
processing.}

\item{grid.size}{The size of each plot grid in nanometers. Default 1000 nm 
for x and y axis, respectively.}

\item{resolution}{ratio of pixel to uM.}

<<<<<<< HEAD:man/plotIndividual_Tracks-methods.Rd
\item{t.interval}{Time interval for image aquisition. Default 0.5 sec 
(500ms).}
=======
\item{t.interval}{Time interval for image aquisition. Default 0.5 sec (500ms).}

\item{sort}{trajectories will be sorted by given parameters, "RT" for dwell time, "Dcoef" for diffusion coefficient, "Rsqare" for R squared of Mean squared Displacement (MSD) fitting.}

\item{legend}{values to be plot as legend on the topright corner of each grid/panel, "RT" for dwell time, "Dcoef" for diffusion coefficient, 
"Both" for both dwell time and diffusion coefficient, "None" for not plotting legend.}

\item{lag}{time intervals for linear regression to calculate diffusion coefficient, default is 2 to 5.}

\item{dt}{max time interval, default is 6.}
>>>>>>> develop:man/plotIndividualTracks-methods.Rd
}
\value{
\itemize{
\item{PDF:} One PDF file with 120 tracks on each page.
}
}
\description{
Plot individual tracks one by one, with grid layout.
<<<<<<< HEAD:man/plotIndividual_Tracks-methods.Rd
Tracks are sorted by their lengths.
}
\details{
Plot individual track/trajectory one by one, and the tracks will 
be layout in 15X8 grids.
=======
             Tracks can be sorted by their lengths or diffusion coefficients.
}
\details{
Plot individual track/trajectory one by one, and the tracks will be layout in 15X8 grids. 
The title of each grid indicates the sequence of the track in the original trackll.
>>>>>>> develop:man/plotIndividualTracks-methods.Rd
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be 
# merged.
folder=system.file('extdata','HSF',package='sojourner')
trackll=createTrackll(folder=folder, input=3)
trackll=filterTrack(trackll,filter=c(7,Inf))
trackll=maskTracks(folder,trackll)
trackll=mergeTracks(folder, trackll)

# Plot individual tracks,
<<<<<<< HEAD:man/plotIndividual_Tracks-methods.Rd
plotIndividual_Tracks(trackll,grid.size=c(1000,1000),resolution=0.107,
                     t.interval=0.5)
=======
plotIndividualTracks(trackll,grid.size=c(1000,1000),resolution=0.107,t.interval=0.5,sort="RT",legend="Both",lag=c(2:5),dt=6)
>>>>>>> develop:man/plotIndividualTracks-methods.Rd
}
