% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualMSD.R
\docType{methods}
\name{plotIndividualMSD}
\alias{plotIndividualMSD}
\title{plotIndividualMSD}
\usage{
plotIndividualMSD(trackll=trackll,dt=6,lag=c(2:5),resolution=0.107,t.interval=0.01,sort="Rsquare",plot.fit=TRUE)
}
\arguments{
\item{trackll}{trajectory list generated by createTrackll() and processing.}

\item{dt}{max time interval, default is 6.}

\item{lag}{time intervals for linear regression to calculate diffusion coefficient, default is 2 to 5.}

\item{resolution}{ratio of pixel to µM.}

\item{t.interval}{Time interval for image aquisition. Default 0.5 sec (500ms).}

\item{sort}{trajectories will be sorted by given parameters, "RT" for dwell time, "Dcoef" for diffusion coefficient, "Rsqare" for R squared of Mean squared Displacement (MSD) fitting.}
}
\value{
\itemize{
\item{PDF:} One PDF file with 120 tracks on each page.
}
}
\description{
Plot the correlation between diffusion coefficient and Rsquare of the MSD fit in a trackll. 
With an option to plot individual MSD and linear regression of all tracks one by one, with grid layout.
             Tracks can be sorted by their lengths, diffusion coefficient or R square of the MSD linear regresson.
}
\details{
Plot individual track/trajectory one by one, and the tracks will be layout in 15X8 grids.
The title of each grid indicates the sequence of the track in the original trackll.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be merged.
folder=system.file("extdata","HSF",package="sojourner")
trackll=createTrackll(interact=F,folder,input=2, cores = 2)
trackll=maskTracks(folder,trackll)
trackll=mergeTracks(folder, trackll)
trackll=filterTrack(trackll,filter=c(7,Inf))

# Plot individual tracks,
plotIndividualMSD(trackll=trackll,dt=6,lag=c(2:5),resolution=0.107,t.interval=0.01,sort="Rsquare",plot.fit=TRUE)
}
