% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualTracks.R
\docType{methods}
\name{plotIndividualTracks}
\alias{plotIndividualTracks}
\title{plotIndividualTracks}
\usage{
plotIndividualTracks(trackll=trackll,grid.size=c(1000,1000),resolution=0.107,t.interval=0.5,sort="RT",legend="Both",lag=c(2:5),dt=6)
}
\arguments{
\item{trackll}{trajectory list generated by createTrackll() and processing.}

\item{grid.size}{The size of each plot grid in nanometers. Default 1000 nm for x and y axis, respectively.}

\item{resolution}{ratio of pixel to µM.}

\item{t.interval}{Time interval for image aquisition. Default 0.5 sec (500ms).}

\item{sort}{trajectories will be sorted by given parameters, "RT" for dwell time, "Dcoef" for diffusion coefficient, "Rsqare" for R squared of Mean squared Displacement (MSD) fitting.}

\item{legend}{values to be plot as legend on the topright corner of each grid/panel, "RT" for dwell time, "Dcoef" for diffusion coefficient, 
"Both" for both dwell time and diffusion coefficient, "None" for not plotting legend.}

\item{lag}{time intervals for linear regression to calculate diffusion coefficient, default is 2 to 5.}

\item{dt}{max time interval, default is 6.}
}
\value{
\itemize{
\item{PDF:} One PDF file with 120 tracks on each page.
}
}
\description{
Plot individual tracks one by one, with grid layout.
             Tracks can be sorted by their lengths or diffusion coefficients.
}
\details{
Plot individual track/trajectory one by one, and the tracks will be layout in 15X8 grids. 
The title of each grid indicates the sequence of the track in the original trackll.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be merged.
folder=system.file("extdata","HSF",package="sojourner")
trackll=createTrackll(interact=F,folder,input=2, cores = 2)
trackll=maskTracks(folder,trackll)
trackll=mergeTracks(folder, trackll)
trackll=filterTrack(trackll,filter=c(7,Inf))

# Plot individual tracks,
plotIndividualTracks(trackll,grid.size=c(1000,1000),resolution=0.107,t.interval=0.5,sort="RT",legend="Both",lag=c(2:5),dt=6)
}
