% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readParticleTracker.R
\docType{methods}
\name{readParticleTracker}
\alias{.readParticleTracker}
\alias{readParticleTracker}
\title{readParticleTracker}
\usage{
readParticleTracker(folder, merge = F, ab.track = F, mask = F, cores = 1, frameRecord = T)

.readParticleTracker(file, interact = F, ab.track = F, frameRecord = F)
}
\arguments{
\item{frameRecord}{Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames).}

\item{folder}{Full path to Diatrack output file.}

\item{merge}{An logical indicate if the output list should be merged into one. Default merge = FALSE, output list is divided by file names.}

\item{mask}{An logical indicate if image mask should be applied to screen tracks. Default False. Note the mask file should have the same name as the Diatrack output txt file with a "_MASK.tif" ending. Users can use plotMask() and plotTrackOverlay() to see the mask and its effect on screening tracks.}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralelled.}
}
\value{
\itemize{
\item{merge = F} Defult. A list of list of data.frames. First level is a list of file names in Diatrack output folder, second level is a list of data.frames from individual output file.
\item{merge = T} A list list of data.frames. First level is the folder name. second level is a list of data.frames from all Diatrack output files merged into one
}
}
\description{
Read output file (tracks/trajectories in csv format) from ParticleTracker (a program of ImageJ plugin MosaicSuit).
}
\details{
The usage of readParticleTracker() is equivalent to ReadDiatrack().
default merge = FALSE, so the researcher can assay variations between files. Keep both output as two level list is for simplicity of downstream analysis.

Note: the folder name should not contain ".", as it is a key character for subsequent indexing of file names.

trackID=fileID.frameID.duration.indexPerFile.indexPerTrackll

This "indexPerFile" is the index within a diatrackFile.

This "indexPerTrackll" is the index within a trackll, which is unique.

The macro used for generating the csv file is also included in ImageJ folder of the package: folder=system.file("extdata","ImageJ",package="smt")
}
\examples{
# reading in tracks
folder=system.file("extdata","ImageJ",package="smt")
trackll=readParticleTracker(folder)
str(trackll,max.level=2)

# masking with image mask
trackll.masked=readParticleTracker(folder=folder,merge=F,mask=T)
str(trackll.masked,1)

# compare the masking effect
plotTrackOverlay(trackll)
plotTrackOverlay(trackll.masked)

# if Nucclear image is available
plotNucTrackOverlay(folder=folder,trackll=trackll)
plotNucTrackOverlay(folder=folder,trackll=trackll.masked)

# plot mask
plotMask(folder=folder)
}

