% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSlimFast.R
\docType{methods}
\name{readSlimFast}
\alias{readSlimFast}
\alias{.readSlimFast}
\title{readSlimFast}
\usage{
readSlimFast(folder, ab.track = F, cores = 1, frameRecord = T)
.readSlimFast(file, interact = F,  ab.track = F, frameRecord = F)
}
\arguments{
\item{file}{Full path to Diatrack .mat session file.}

\item{interact}{Open menu to interactively choose file.}

\item{ab.track}{Use absolute coordinates for tracks.}

\item{frameRecord}{Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames).}

\item{folder}{Full path to SlimFast .txt files output folder.}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralleled.}
}
\description{
take in a SlimFast .txt session file as input, along with several other user-configurable parameters and output options, to return a track list of all the trajectories
}
\details{
The naming scheme for each track is as follows:

[Last five characters of the file name].[Start frame #].[Length].[Track #]

(Note: The last five characters of the file name, excluding the extension, cannot contain ".")
}
\examples{
#Basic function call of .readSlimFast
#txtfolder=system.file("extdata","HTZ1",package="sojourner")
#trackll <- readSlimFast(folder = txtfolder, cores = 1)

#Basic function call of .readSlimFast
# trackl <- .readSlimFast(interact = TRUE)

}
