% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMate.R
\docType{methods}
\name{readTrackMate}
\alias{readTrackMate}
\alias{.readTrackMate}
\title{readTrackMate}
\usage{
readTrackMate(folder, ab.track = F, cores = 1, frameRecord = T)

.readTrackMate(file, interact = F, ab.track = F, frameRecord = F)
}
\arguments{
\item{file}{Full path to Diatrack .mat session file.}

\item{interact}{Open menu to interactively choose file.}

\item{ab.track}{Use absolute coordinates for tracks.}

\item{frameRecord}{Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames).}

\item{folder}{Full path to ImageJ .csv files output folder.}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralleled.}
}
\description{
Read output file (tracks/trajectories in csv format) from TrackMate (a program of ImageJ plugin).
}
\details{
Note: the folder name should not contain ".", as it is a key character for subsequent indexing of file names.

trackID=fileID.frameID.duration.indexPerFile.indexPerTrackll

This "indexPerFile" is the index within a diatrackFile.

This "indexPerTrackll" is the index within a trackll, which is unique.

The usage of readTrackMate() is similar to ReadDiatrack(), please reference for more details.
}
\examples{
## Not run:
trackMate.file.url="https://drive.google.com/file/d/1abYagKpu1yea_7FM8tJ7nkQdOktFyS6j/view?usp=sharing"
# with the upgrade of cloud protocol, this command does not work anymore, 
# download and save file to a location on local computer 
# download.file(trackMate.file.url, "~/test/test.csv")
# dir.create("~/test/")
# folder=~/test/"
# trackll=readTrackMate(folder)
# str(trackll,max.level=2)

}
