% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentTrack.R
\docType{methods}
\name{segmentTrackll}
\alias{segmentTrackll}
\alias{segmentTrack}
\title{segmentTrackll}
\usage{
segmentTrack(track, seg)
}
\description{
generate track segments of desired length with given trackll variable.
segment length of 6 or greater is needed. Tracks that are shorter than the
indicated seg value will be sorted out.
}
