% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smt.R
\docType{package}
\name{smt}
\alias{smt}
\alias{smt-package}
\title{smt-simple parameters for single molecule tracking analysis}
\usage{
smt()
}
\description{
simple analysis on single molecule tracking data using parameters based on mean square displacement (MSD).
}
\details{
smt provide a simple analysis on single molecule tracking data using parameters based on mean square displacement (MSD). Currently includes:
- duration of the tracks (dwellTime),

- square displacement (squareDisp),

- mean square displacement as a function of time (msd),

- diffusion coefficient (Dcoef) and

- emperical cumulative distribution function (eCDF) of MSD over time.
}

