% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sojourner.R
\docType{package}
\name{sojourner}
\alias{sojourner}
\alias{sojournerGUI}
\title{sojourner-simple parameters for single molecule tracking analysis}
\usage{
sojourner()
sojournerGUI()
}
\description{
simple analysis on single molecule tracking data using parameters based on mean square displacement (MSD).
}
\details{
sojourner provide a simple analysis on single molecule tracking data using parameters based on mean square displacement (MSD). Currently includes:
- duration of the tracks (dwellTime),

- square displacement (squareDisp),

- mean square displacement as a function of time (msd),

- diffusion coefficient (Dcoef) and

- emperical cumulative distribution function (eCDF) of MSD over time.
}
