% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sojournerGUI.R
\docType{methods}
\name{sojournerGUI}
\alias{sojournerGUI}
\title{sojournerGUI}
\usage{
sojournerGUI()
}
\value{
A graphical user interface for sojourner.
}
\description{
Graphical user interface (GUI) for sojourner.
}
\details{
The GUI version has the most frequently used functions but not all command
line ones. Current version includes: MSD (mean square displacement),
Dcoef(diffusion coefficient), CDF(cumulative distribution function), and
RT(residence time) etc.
}
\examples{
# not run
# library(sojourner)
# sojournerGUI()

}
