% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareDisp.R
\docType{methods}
\name{squareDisp}
\alias{squareDisp}
\alias{squareDispCpp}
\title{squareDisp}
\usage{
squareDisp(track,dt=1,resolution=0.107)
squareDispCpp(track,dt=1,resolution=0.107)
}
\arguments{
\item{track}{track dataframe with x and y coordinates.}

\item{dt}{time step size(in frames).}

\item{resolution}{resolution value, default is 0.107.}
}
\value{
\itemize{
\item{list of square displacements(dx^2 + dy^2) for varying dt
values from 1 to dt}
}
}
\description{
calculate square displacement of a track/trajectory as a function of 
time/step. data.frame has two column, x and y
also calculate dx, dy bivariate
squareDispCpp is the cpp version of squareDisp
}
\examples{
folder1=system.file('extdata','SWR1',package='sojourner')
folder2=system.file('extdata','HTZ1',package='sojourner')
trackll=compareFolder(folders=c(folder1,folder2), input=3)
#use default filter with min=7
filtered.trackll=filterTrack(trackll)
track.dt=squareDisp(filtered.trackll[[1]][[1]],dt=6)
}
