% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{.boot.se.onecomp}
\title{bootstrap}
\usage{
bootstrap(normalFit,n.reps=100)
  
  plotBootstrap(d.boot,alpha=1/2) -- not available in current version
}
\arguments{
\item{n.reps}{number of replicates for bootstrapping}

\item{normalFit}{output from fitNormDistr}

\item{d.boot}{bootstrapped data, or the output from bootstrap}

\item{alpha}{transparency adjustment, between 0 to 1.}
}
\value{
\itemize{ 
List of length 2 containing:
\item{Fit}
results from fitNormDistr that was used as the input for the function
\item{Bootstraps}
results from bootstrapping. mean values of each samples,
standard error for the mean, lambda(proportion)
 }
}
\description{
Bootstrap confidience intervals with standard errors. bootstrap resamples dataset (e.g. diffusion coefficients) to calculate confidience intervals for a statistic measure of dataset.
}
\details{
A wrapper of boot::boot and mixtools::boot.se adapted for data format in sojourner package. 
Also returns stderr information by running boot.se from mixtools  
and an additional method for one-component distribution calculates the stderr separately.
For multi-component distributions, the boot.se function from the mixtools package was used.
For single-component distributions, a separate function was used to calculate the stderr and confidence interval.
}
\examples{
# read in using readDiatrack
folder=system.file("extdata","SWR1",package="sojourner")
trackll=readDiatrack(folder)

# filter track based length
trackll.flt=filterTrack(trackll,filter=c(min=5,max=Inf))
MSD=msd(trackll.flt,dt=6,summarize=FALSE,plot=TRUE)
dcoef=Dcoef(MSD=MSD,method="static",plot=FALSE)
# fit the dcoef result
normalFit=fitNormDistr(dcoef)

# bootstrap new datasets
d.boot=bootstrap(normalFit)
# manually set the number of bootstrap samples to 50
d.boot=bootstrap(normalFit, n.reps=50)

}
