% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTrackll.R
\docType{methods}
\name{combineTrackll}
\alias{combineTrackll}
\title{combineTrackll}
\usage{
combineTrackll(total=2,merged=T)
}
\arguments{
\item{total}{The number of tracklls to combine.}

\item{merged}{An Logical indicate if the tracklls to combine are merged or not.}
}
\value{
\itemize{
\item{trackll:} combined trackll.
}
}
\description{
Combine multiple tracklls into one trackll.
}
\details{
Combine multiple track listw (tracklls) from multiple folders into one trackll, i.e. combining track infomation
         from files in multiple folders (replicates) together as if they are in one folder. The tracklls 
         either merged or un-merged.
         
         Users will be prompted to input the name of each trackll to combine.
         If the combined tracklls are merged, users will also be prompted to input a combined attribute name for the combined trackll.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, merge tracks from multiple files.
folder1=system.file("extdata","SWR1",package="smt")
trackll1=createTrackll(interact=F,folder,input=1)
trackll1=maskTracks(folder,trackll)
trackll1=mergeTracks(folder,trackll)

folder2=system.file("extdata","SWR1_2",package="smt")
trackll2=createTrackll(interact=F,folder,input=1)
trackll2=maskTracks(folder,trackll)
trackll2=mergeTracks(folder,trackll)

# Combine the tracklls together, input trackll names when prompted,
trackll=combineTrackll(total=2,merged=T)
}
