% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTrackll.R
\docType{methods}
\name{combineTrackll}
\alias{combineTrackll}
\title{combineTrackll}
\usage{
combineTrackll(trackll,name='combined trackll',merged=TRUE)
}
\arguments{
\item{trackll}{The tracklls to be combined together.}

\item{name}{a character string given to set the 'names' attribute
for the combined trackll}

\item{merged}{An Logical indicate if the tracklls to combine are 
merged or not.}
}
\value{
\itemize{
\item{trackll:} combined trackll.
}
}
\description{
Combine multiple tracklls into one trackll.
}
\details{
Combine multiple track lists (tracklls) from multiple folders
into one trackll, i.e. combining track infomation from files in multiple 
folders (replicates) together as if they are in one folder. The tracklls
can be either merged or un-merged.

The name argument sets the 'names' attribute for the combined trackll, 
which will be used in the same way as the folder names for the original
tracklls, e.g., displayed as legend when plotting Dcoef or MSD for 
the combined trackll.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, merge tracks from multiple files.
folder1=system.file('extdata','HSF',package='sojourner')
trackll1=createTrackll(folder1,input=3, cores = 2)
trackll1=maskTracks(folder1,trackll1)
trackll1=mergeTracks(folder1,trackll1)

folder2=system.file('extdata','HSF_2',package='sojourner')
trackll2=createTrackll(folder2,input=2, cores = 2)
trackll2=maskTracks(folder2,trackll2)
trackll2=mergeTracks(folder2,trackll2)

# Combine the tracklls together, input trackll names when prompted,
trackll=combineTrackll(trackll=c(trackll1,trackll2),merged=TRUE)
}
