% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplacementCDF.R
\docType{methods}
\name{displacementCDF}
\alias{displacementCDF}
\alias{displacement.track}
\alias{displacement_trackll}
\title{displacementCDF}
\usage{
displacementCDF(trackll,dt=1,resolution=0.107,plot=FALSE,output=FALSE,
bivar=FALSE)
displacement_trackll(trackll,dt=1,resolution=0.107,bivar=FALSE)
}
\arguments{
\item{dt}{Time intervals.}

\item{resolution}{ratio of pixel to uM.}

\item{bivar}{bivar=FALSE, view displacement r as single variable; 
bivar=TRUE, view x,y as bivarate. Default value F.}

\item{trackll}{a list of track lists.}

\item{plot}{An logical indicate if plot should be generated. See Values for
detail.}

\item{output}{An logical indicate if output should be generated. See Values
for detail.}
}
\value{
\itemize{
\item{ list of "stepwise.displacement" and "CDF.displacement",} {A list of
stepwise.displacement" and "CDF.displacement". the name of the list is the
track folder name.}

\item{Output file,} {Displacement of individual trajectoreis at specified
dt. The output file is for user to plot in other applications. The column
"UniqueDisplacement" is the x axis, and column "CDF" is the y axis for a 
CDF plot. The distribution of "UniqueDisplacement" is the density plot. }


\item{CDF plot,} {CDF plot of displacement for individual files. 
}
}
}
\description{
calculate cumulative distribution function of all displacement
for individual trajectories.
}
\details{
The cumulative radial distribution function, P(r, i*dt), is the
probability of finding the diffusing particle within a radius r from the
origin at time lag i*dt:

P(r,i*dt) = 1 - e^(-r^2/4*D*(i*dt))

the CDF and UniqueDisplacement in the output file is corresponding to P and
r in this formula. If intend to generate the CDF plot from the output file,
the CDF and UniqueDisplacement is corresponding to the y and x values in 
the CDF output plot.
}
\examples{
folder1=system.file("extdata","SWR1",package="sojourner")
folder2=system.file("extdata","HTZ1",package="sojourner")
trackll=compareFolder(folders=c(folder1,folder2), input=3)
displacementCDF(trackll,dt=1,plot=TRUE)
}
