% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTrack.R
\docType{methods}
\name{filterTrack}
\alias{filterTrack}
\alias{trimTrack}
\alias{trackLength}
\title{filterTrack}
\usage{
filterTrack(trackll,filter=c(min=7,max=Inf))
trimTrack(trackll,trimmer=c(min=1,max=32))
trackLength(trackll)
}
\arguments{
\item{trackll}{a list of track lists.}

\item{filter}{range of possible track lengths to keep}

\item{trimmer}{range of track lengths allowed in output, otherwise trimmed.}
}
\value{
\itemize{
\item{trackll} filtered or trimmed tracks.
\item{len} list of track lengths.
}
}
\description{
methods for filter and trim tracks based on track length.
}
\details{
filterTrack() is used to filter out tracks that has length within 
a specified range (default 7~Inf). On the other hand, despite the lengths 
of tracks, trimTrack() is used to trim /cutoff all tracks to a specified 
range (default 1~32).
}
\examples{
folder=system.file('extdata','SWR1',package='sojourner')
trackll=createTrackll(folder=folder, input=3)

trackll.filter=filterTrack(trackll,filter=c(7,Inf))
trackll.trim=trimTrack(trackll,trimmer=c(1,20))

# see the min and max length of the trackll
# trackLength() is a helper function output track length of trackll
lapply(trackLength(trackll),min)
lapply(trackLength(trackll.filter),min)

lapply(trackLength(trackll),max)
lapply(trackLength(trackll.trim),max)

}
