% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitResidenceTime.R
\docType{methods}
\name{fitRT}
\alias{fitRT}
\alias{.one.comp.fit.rt}
\title{Fitting Residence time (1-CDF)}
\usage{
fitRT(trackll=NULL,x.max=30,N.min=1.5,t.interval=0.5,maxiter.search=1e3,
maxiter.optim=1e3,k.ns=FALSE)
}
\arguments{
\item{t.interval}{time interval for image aquisition. Default 0.5 sec.}

\item{maxiter.optim}{A positive integer specifying the maximum number of iterations allowed for nls.}

\item{trackll}{trajectory list generated by createTrackll() and processing. if NULL, user will be prompted to enter the trackll name.}

\item{x.max}{The maximum range of X axis, i.e. time, for the output plot. Default 30 sec.}

\item{N.min}{The minimal duration of trajectory length in the unit of second, trajectories shorter than N.min will be filtered out. Default 1.5 sec.}

\item{k.ns}{Logical indicate or numeric used to control fitting.}

\item{maxiter.search}{A positive integer specifying the maximum number of iterations allowed for nls.}
}
\value{
\itemize{
\item{On the Console output:} Result of both one and two-component fit and parameters of goodness of the fit.
\item{Plot:} fitting curves will be plotted over the raw data, with the number of tracks and result of two-component fitting shown in the plot area.
}
}
\description{
Caclulate average residence time for trajecotries by  
fitting 1-CDF (survival distribution).
}
\details{
Calculating average residence time of particles by fitting 1-CDF (survival distribution) of its trajectorys.

Input track list (trackll) should be processed, merged trackll, the fitting will start right away after running the function. 
The maximum time range to be plotted can be set using x.max, this will not change the fitting result. The result of two-component fit will be shown on the plot as well as in the console.
Fitting result is determined by the input trackll, and N.min (filtering of the trackll).
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, merge tracks from multiple files.
folder=system.file("extdata","HSF",package="sojourner")
trackll=createTrackll(folder=folder, input=3)
trackll=maskTracks(folder,trackll)
trackll=mergeTracks(folder,trackll)

# Fit the residence time of trackll
fitRT(trackll=trackll,x.max=30,N.min=1.5,t.interval=0.5)
}
