% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\docType{methods}
\name{getCI}
\alias{getCI}
\alias{.get.seRange}
\title{getCI}
\usage{
getCI(bootstrap.result,confidence=0.95,output=FALSE)
}
\arguments{
\item{confidence}{the level of confidence that is used to calculate the
confidence interval.}

\item{bootstrap.result}{diffusion coefficient calculated from Dcoef().}

\item{output}{Logical indicate if output file should be generated.}
}
\value{
list of items, each of which will contain for each distribution component:
\describe{
\item{Estimate}{Mean estimated from sample}
\item{CI lower}{Lower bound of the confidence interval}
\item{CI upper}{Upper bound of the confidence interval}
\item{Std. Error}{Std. Error for given data}
}
}
\description{
given a dcoef result, return Confidence Interval
information(mean, bounds, std err) for diffusion coefficient and the
proportions of each components.
}
\details{
Supplied with a bootstrap output, it calculates the confidence
range. The t-distribution/critical-t was used to calculate the Confidence
Interval.
}
\examples{
# compare folders
folder1=system.file("extdata","SWR1",package="sojourner")
folder2=system.file("extdata","HTZ1",package="sojourner")
trackll=compareFolder(c(folder1,folder2))
#get msd
MSD=msd(trackll=trackll)
#run Dcoef()
dcoef=Dcoef(MSD,dt=6,plot=TRUE,output=FALSE)
#fit the dcoef result
normalFit=fitNormDistr(dcoef)
# perform bootstrapping for this dcoef result
d.boot = bootstrap(normalFit, n.reps=100)
# get confidence intervals for this dcoef result which contains data from two different folders
a=getCI(d.boot)
# to manually set confidence to 80\%
b=getCI(d.boot, confidence=0.8, output=FALSE)

}
