% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmTrackll.R
\docType{methods}
\name{hmmTrackll}
\alias{hmmTrackll}
\title{hmmTrackll}
\usage{
hmmTrackll(trackll, t.interval=0.01)
}
\arguments{
\item{trackll}{list of trajectorise created by crreateTrackll().}

\item{t.interval}{time interval.}
}
\value{
Hidden Markov Model fitted trackll
}
\description{
Convert trackll to a (bivariate) displacement format for Hidden 
Markov Model fitting.
}
\details{
dx, displacement on x.

dy, displacement on y. 

time, lenght of time a particle has being traveling. 

trackStepIndex, steps index within a track.
}
\examples{
folder=system.file('extdata','SWR1',package='sojourner')
trackll=createTrackll(folder=folder, input=3)
hTrackll=hmmTrackll(trackll)

}
