% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maskTracks.R
\docType{methods}
\name{maskTracks}
\alias{maskPoint}
\alias{maskTracks}
\title{maskTracks}
\usage{
maskTracks(folder, trackll)
}
\arguments{
\item{folder}{Full path to the output files.}

\item{trackll}{A list of track lists.}
}
\description{
apply binary image masks to lists of track lists
}
\details{
IMPORTANT: It will take an extremely long time to mask unfiltered data. Filter first using filterTrack(trackll,filter=c(min=7,max=Inf)), then mask using maskTracks(folder, trackll)!
Note the mask file should have the same name as the output files with a "_MASK.tif" ending.
If there are more mask files than trackll, masking will fail. If there are less mask files, trackls without masks will be deleted.
Users can use plotMask() and plotTrackOverlay() to see the mask and its effect on screening tracks.
}
\examples{
#Basic masking with folder path with image masks
folder = "/FOLDERPATH"
trackll.masked <- maskTracks(folder = folder, trackll = trackll)

#Compare the masking effect
plotTrackOverlay(trackll)
plotTrackOverlay(trackll.masked)

#Plot mask
mask.list=list.files(path=folder,pattern="_MASK.tif",full.names=T)
plotMask(mask.file=mask.list[[1]])

#If Nuclear image is available
plotNucTrackOverlay(folder=folder,trackll=trackll)
plotNucTrackOverlay(folder=folder,trackll=trackll.masked)

#Plot mask
plotMask(folder=folder)
}

