% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividual_Tracks.R
\docType{methods}
\name{plotIndividual_Tracks}
\alias{plotIndividual_Tracks}
\title{plotIndividual_Tracks}
\usage{
plotIndividual_Tracks(trackll=trackll,grid.size=c(1000,1000),
resolution=0.107,t.interval=0.5)
}
\arguments{
\item{trackll}{trajectory list generated by createTrackll() and 
processing.}

\item{grid.size}{The size of each plot grid in nanometers. Default 1000 nm 
for x and y axis, respectively.}

\item{resolution}{ratio of pixel to uM.}

\item{t.interval}{Time interval for image aquisition. Default 0.5 sec 
(500ms).}
}
\value{
\itemize{
\item{PDF:} One PDF file with 120 tracks on each page.
}
}
\description{
Plot individual tracks one by one, with grid layout.
Tracks are sorted by their lengths.
}
\details{
Plot individual track/trajectory one by one, and the tracks will 
be layout in 15X8 grids.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be 
# merged.
folder=system.file('extdata','HSF',package='sojourner')
trackll=createTrackll(folder=folder, input=3)
trackll=filterTrack(trackll,filter=c(7,Inf))
trackll=maskTracks(folder,trackll)
trackll=mergeTracks(folder, trackll)

# Plot individual tracks,
plotIndividual_Tracks(trackll,grid.size=c(1000,1000),resolution=0.107,
                     t.interval=0.5)
}
