% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocalizations.Density.R
\docType{methods}
\name{plotLocalizations.Density}
\alias{plotLocalizations.Density}
\alias{.plotLocalizations.Density}
\title{plotLocalizations.Density}
\usage{
plotLocalizations.Density(trackll=trackll,scale=256,r=125,file.No=0,point.scale=0.15)
}
\arguments{
\item{trackll}{trajectory list generated by createTrackll() and processing. if NULL, user will be prompted to enter the trackll name.}

\item{scale}{The pixel scale of image data.}

\item{r}{Radius within each molecule to calculate density.}

\item{file.No}{Select file(s) in the folder to plot. Default 0 for plotting all files in the folder.}

\item{point.scale}{Size of the dots representing the molecules.}
}
\value{
\itemize{
\item{PDF:} One PDF file with one plot on each page.
}
}
\description{
Plot localization map of molecules from a list of files in a folder with color coded 
             by local density of each molecule.
}
\details{
Plot localization map of molecules from a list of files in a folder with color coded 
             by local density of each molecule. The localization of molecule is considered as the first position of its track. 
             
         Upon running of the function, users will be prompted to input the name of the track list (trackll).
         Input un-merged trackll and the plotting will start.
         The local density of each molecule is calculated by counting the number of molecules within a given radius
         around the position of the molecule. The higher the number, the higher the local density.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be merged.
folder=system.file("extdata","HSF",package="smt")
trackll=createTrackll(interact=F,folder,input=2)
trackll=maskTracks(folder,trackll)

# Plot localization map,
plotLocalizations.Density(trackll=trackll,scale=128,r=125,file.No=0,point.scale=0.3)
plotLocalizations.Density(trackll=trackll,scale=128,r=125,file.No=2,point.scale=1)

}
