% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrackOverlay.Dcoef.R
\docType{methods}
\name{plotTrackOverlay.Dcoef}
\alias{plotTrackOverlay.Dcoef}
\alias{.plotTrackOverlay.Dcoef}
\title{plotTrackOverlay.Dcoef}
\usage{
plotTrackOverlay.Dcoef(trackll=trackll,scale=128,dt=6,filter=c(min=7,max=Inf),resolution=0.107,rsquare=0.8,
               t.interval=0.01,Dcoef.range=c(-6,2),color=c("blue", "white", "red"),
               folder=NULL,file.No=0,line.width=0.1)
}
\arguments{
\item{trackll}{trajectory list generated by createTrackll() and processing. if NULL, user will be prompted to enter the trackll name.}

\item{scale}{The pixel scale of image data.}

\item{dt}{Time inbervals.}

\item{filter}{Filter the tracks by step/frame number (length). Only tracks pass through filter will be selected.}

\item{resolution}{ratio of pixel to µM.}

\item{rsquare}{R square filter on Dcoef results. Default to be 0.8. Set value to 0 if rsquare filter is not desired.}

\item{t.interval}{Time interval for image aquisition. Default 0.01 sec (10ms).}

\item{Dcoef.range}{Select tracks whose Dcoef is within the range. Other tracks will not be plotted.}

\item{color}{colors to interpolate to build the color gradient/ramp.}

\item{folder}{Path to provide nuclear glow backround. Images should end with "_Nuclei.tif". Default is NULL and not adding background.}

\item{file.No}{Select file(s) in the folder to plot. Default 0 for plotting all files in the folder.}

\item{line.width}{Line width for plotted tracks, default is 0.1.}
}
\value{
\itemize{
\item{PDF:} One PDF file with one plot on each page.
}
}
\description{
Plot track/trajectory overlays from a list of files in a folder with color coded 
             by the Diffusion Coefficient (Dcoef) of each track/trajectory.
}
\details{
Plot track/trajectory overlays from a list of files in a folder with color coded 
         by the Diffusion Coefficient (Dcoef) of each track/trajectory. The range of Dcoef can be defined. Nuclei backgrounds can be added to the plots.
         Tracks from each file will be plot in one plot, i.e., one plot per file. Colors are flexible.
             
         Upon running of the function, users will be prompted to input the name of the track list (trackll).
         Input un-merged trackll and the plotting will start. 
         The Dcoef is calculated by "static" method, which stabilizes the number of time lags used for fitting using time lag 2~ 5 despite the total time lags measured.
}
\examples{

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be merged.
folder=system.file("extdata","HSF",package="sojourner")
trackll=createTrackll(interact=F,folder,input=2, cores = 2)
trackll=maskTracks(folder,trackll)

# Plot track overlays,
plotTrackOverlay.Dcoef(trackll=trackll,scale=128,dt=6,filter=c(min=7,max=Inf),resolution=0.107,rsquare=0.8,
               t.interval=0.01,Dcoef.range=c(-6,2),color=c("blue", "white", "red"),
               folder=NULL,file.No=0,line.width=0.1)
plotTrackOverlay.Dcoef(trackll=trackll,color=c("red", "yellow"),folder=folder,file.No=0)
plotTrackOverlay.Dcoef(trackll=trackll,color=c("red", "yellow"),folder=folder,file.No=c(1,2))
}
