% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readDiatrack.R
\docType{methods}
\name{readDiatrack}
\alias{.readDiatrack}
\alias{readDiatrack}
\title{readDiatrack}
\usage{
readDiatrack(folder, ab.track = F, cores = 1, frameRecord = T)

.readDiatrack(file, interact = F, ab.track = F, frameRecord = F)

maskTracks(folder, trackll)
}
\arguments{
\item{file}{Full path to Diatrack .mat session file.}

\item{interact}{Open menu to interactively choose file.}

\item{ab.track}{Use absolute coordinates for tracks.}

\item{frameRecord}{Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames).}

\item{folder}{Full path to Diatrack .txt session files output folder.}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralleled.}
}
\description{
read output file (tracks/trajecotries) from Diatrack.
}
\details{
Note: the folder name should not contain ".", as it is a key charactero for subsequent indexing of file names.

the absolute coordinates trajectory has moved

trackID=fileID.frameID.duration.indexPerFile.indexPerTrackll

This "indexPerFile" is the index within a diatrackFile.

This "indexPerTrackll" is the index within a trackll, which is unique.
}
\examples{
folder=system.file("extdata","SWR1",package="smt")
trackll=readDiatrack(folder)
str(trackll,max.level=2)
}

