% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSlimFast.R
\docType{methods}
\name{readSlimFast}
\alias{readSlimFast}
\alias{.readSlimFast}
\title{readSlimFast}
\usage{
readSlimFa(folder, merge = F, ab.track = F, mask = F, cores = 1, frameRecord = T)
.readSlimFast(file, interact = F,  ab.track = F, frameRecord = F)
}
\arguments{
\item{file}{Full path to SlimFast .txt session file.}

\item{interact}{Open menu to interactively choose file.}

\item{ab.track}{Use absolute coordinates for tracks.}

\item{frameRecord}{Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames).}

\item{folder}{Full path Diatrack .mat session files output folder.}

\item{merge}{Indicate if the output list should be merged into one- output list is divided by file names otherwise.}

\item{mask}{Indicate if image mask should be applied to screen tracks (Note: the mask file should have the same name as the Diatrack output txt file with a "_MASK.tif" ending. Users can use plotMask() and plotTrackOverlay() to see the mask and its effect on screening tracks).}

\item{cores}{Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralleled.}
}
\description{
take in a SlimFast .txt session file as input, along with several other user-configurable parameters and output options, to return a track list of all the trajectories
}
\details{
The naming scheme for each track is as follows:

[Last five characters of the file name].[Start frame #].[Length].[Track #]

(Note: The last five characters of the file name, excluding the extension, cannot contain “.”)
}
\examples{
#Basic function call of .readSlimFast
trackll <- readSlimFast(folder = /FILEPATH/, cores = 2)

#Basic function call of .readSlimFast
trackl <- .readSlimFast(interact = T)

}
